Removes GA plugin tight coupling
This commit removes the direct integration of the angulartics GA plugin
in favour of the parent project being able to select which angulartics
plugins should be used.

The features here are still weighted towards using the GA plugin with
support for indexed custom dimensions (even the 'custom dimensions'
terminology is borrowed from GA), and the README reflects this with
several nudges towards using the GA plugin.

An improvement in the future may be to further seperate concerns by
having translator plugins for at-ng-analytics that are run through to
generate what exactly gets sent to angulartics in the userProperties
map.